refactor(test): rename pkg 'test' to 'testsupport'
also, rename `test/resources/all-good.*` => `testsupport/resources/fake-api-server.*`
since these files are used to test the fake API server itself
