--- test/resources/replicasets.yaml
@@ -1,50 +1,0 @@
-apiVersion: apps/v1
-kind: ReplicaSet
-metadata:
-  namespace: test
-  name: rs-1
-  labels:
-    app: rs-1
-    more: cookies
-
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: rs-1
-  template:
-    metadata:
-      labels:
-        app: rs-1
-    spec:
-      containers:
-      - name: default
-        image: caddy:2
-        imagePullPolicy: Always
-      serviceAccount: default
-      serviceAccountName: default
----      
-apiVersion: apps/v1
-kind: ReplicaSet
-metadata:
-  namespace: test
-  name: rs-2
-  labels:
-    app: rs-2
-    more: cookies
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: rs-2
-  template:
-    metadata:
-      labels:
-        app: rs-2
-    spec:
-      containers:
-      - name: default
-        image: caddy:2
-        imagePullPolicy: Always
-      serviceAccount: default
-      serviceAccountName: default
