+++ testsupport/fake_api_server.go
@@ -1,4 +1,4 @@
-package test
+package testsupport
 
 import (
 	"bufio"
