+++ testsupport/resources/deployment-zero-replica.yaml
