+++ testsupport/rest_config.go
@@ -1,4 +1,4 @@
-package test
+package testsupport
 
 import (
 	corev1 "k8s.io/api/core/v1"
